refactor: 整理代码并优化初始化与反馈组件
1.  移除多语言文件中无用的app启动阶段配置
2.  重构选中武器条组件，提取为独立文件
3.  优化初始化加载覆盖层和反馈组件，支持自定义反馈链接
4.  修复武器卡片按钮样式，替换为标准Button组件
