+++ src/components/essence/dungeon-card.tsx
@@ -163,7 +163,7 @@ const WeaponThumbnail = memo(function WeaponThumbnail({
     pointerStartRef.current = null
     if (longPressTriggeredRef.current) {
       setOpen(false)
-      longPressTriggeredRef.current = false
+      // Do NOT reset longPressTriggeredRef here — let handleToggle consume it
     }
   }, [clearLongPress])
 
@@ -182,8 +182,9 @@ const WeaponThumbnail = memo(function WeaponThumbnail({
   const weaponName = (weapon.id?.startsWith('custom-') || weapon.id?.startsWith('preview:')) ? weapon.name : (t(`weapons.${weapon.id}`) ?? weapon.name)
 
   const thumb = (
-    <button
-      type="button"
+    <Button
+      variant="ghost"
+      size="icon"
       ref={triggerRef}
       onClick={handleToggle}
       onPointerDown={isMobile && enableTooltip ? handlePointerDown : undefined}
@@ -216,7 +217,7 @@ const WeaponThumbnail = memo(function WeaponThumbnail({
           <Check className="size-2.5 text-black" strokeWidth={3} />
         </div>
       )}
-    </button>
+    </Button>
   )
 
   if (!enableTooltip) return thumb
@@ -273,8 +274,9 @@ const WeaponRow = memo(function WeaponRow({
       )}
     >
       <div className="flex items-center gap-2 min-w-0 shrink-0">
-        <button
-          type="button"
+        <Button
+          variant="ghost"
+          size="icon"
           onClick={() => toggleWeapon(weapon.id)}
           className="relative size-10 rounded shadow-[0_0_0_1px_rgba(0,0,0,0.08)] bg-muted/30 flex-shrink-0 overflow-hidden bg-[url(/images/item-frame-bg.png)] bg-cover bg-center cursor-pointer">
           {(() => { const s = weaponImageSrc(weapon.id); if (!s) return <span className="absolute inset-0 flex items-center justify-center text-lg font-semibold text-white/40">{weapon.name?.charAt(0) ?? '?'}</span>; return <Image src={s} alt={weaponName} fill className="object-cover z-10" unoptimized /> })()}
@@ -286,7 +288,7 @@ const WeaponRow = memo(function WeaponRow({
             className="absolute -inset-x-px bottom-0 z-20 w-[calc(100%+2px)] max-w-none object-cover object-bottom pointer-events-none"
             unoptimized
           />
-        </button>
+        </Button>
         <span
           className={cn(
             'font-medium min-w-0 truncate w-28 flex-shrink-0',
